refactor: Update routing configuration for Vercel
