+++ src/pages/private/ChildrenPage.tsx
@@ -1,12 +1,14 @@
 import { useState } from "react";
-import { Button } from "../../../../components/common/Button";
-import { Card } from "../../../../components/common/Card";
-import { Container } from "../../../../components/common/Container";
-import { Navbar } from "../../../../components/navbars/Navbar";
-import { ShowComponent } from "../../../../components/utils/ShowComponent";
+import { Button } from "../../components/common/Button";
+import { Card } from "../../components/common/Card";
+import { Container } from "../../components/common/Container";
+import { Navbar } from "../../components/navbars/Navbar";
+import { ShowComponent } from "../../components/utils/ShowComponent";
+import { useNavigate } from "react-router-dom";
 
 export const ChildrenPage = () => {
   const [show, setShow] = useState(false);
+  const navigate = useNavigate();
   return (
     <>
       {/* Aquí se puede agregar cualquier contenido que se desee */}
@@ -49,12 +51,17 @@ export const ChildrenPage = () => {
           <p>Este es el contenido de la card 6</p>
         </Card>
       </Container>
-      <Container columns={2} gap={2}>
+      <Container columns={2} gap={4}>
         <Card>
           <h1>Card 1</h1>
           <p>Este es el contenido de la card 1</p>
           <Button title="Click me" onClick={() => setShow(!show)} />
         </Card>
+        <Card>
+          <h1>Card 2</h1>
+          <p>Este boton te lleva a productos</p>
+          <Button title="Productos" onClick={() => navigate("/products")} />
+      </Card>
       </Container>
         {/* Otro ejemplo de uso de un componente que recibe children */}
       <ShowComponent shouldShow={show}>
